AccountManagerPlugin: Rework the 'lost password' procedure, refs #816.
Password reset is less intrusive now, not altering the current password
before a successful login using it.  The temporary password is stored in a
separate SessionStore (sharing configuration with any other SessionStore) and
merely checked as a fallback, if the regular authentication has failed.
On authentication success with the old password any temporary password is
deleted to prevent abuse of the 'lost password' procedure by others.

git-svn-id: http://trac-hacks.org/svn/accountmanagerplugin/trunk@10264 7322e99d-02ea-0310-aa39-e9a107903beb