AccountManagerPlugin: Restore `0.11` compatibility, refs #8963.
Lift unnecessary restriction regarding the `secure_cookies` option, a feature
of Trac 0.11.2 and later, to restrict cookies to HTTPS connections only.
And again I refrain from the tempting idea to fix it in !AcctMgr itself.
The cookie creation isn't totally independed of Trac, and I'm not an expert
to spot possible side-effects of such a move.

git-svn-id: http://trac-hacks.org/svn/accountmanagerplugin/trunk@10428 7322e99d-02ea-0310-aa39-e9a107903beb