AccountManagerPlugin: Add flexible date/time rendering for user lists, refs #6616 and #9852.
Now the time stamps are combined with a relative time interval hint (tool-tip).
This is an enhancement to the user admin panel too.

Support for bleeding-edge user configurable time in Trac 0.13 is accompanied
here by a fallback for Trac 0.11 and 0.12, that looks great and is worth a lot
of the effort put into this rather complicated fallback code.

git-svn-id: http://trac-hacks.org/svn/accountmanagerplugin/trunk@11346 7322e99d-02ea-0310-aa39-e9a107903beb