AccountManagerPlugin: Avoid a `TracError` on logout, refs #10276.
Logging out from 'account' session/user preferences is a forseeable operation,
so it should better not raise an error.  Thanks to Scott for reporting it.
Now, redirect unauthenticated requests to general preferences should be safe.
I've removed some code, that was meant to cope with this situation beforehand,
but it seems to never be reached before Trac decided to error out.

git-svn-id: http://trac-hacks.org/svn/accountmanagerplugin/trunk@11938 7322e99d-02ea-0310-aa39-e9a107903beb