AccountManagerPlugin: Allow filtering of account list in user admin panel by account status, refs #9946 and #10682.
The filter implementation of Trac's query module was used as a guide.
We won't care for banned accounts by default, just active ones and these
marked as pending for approval, on registration time and via the email
verification process, if enabled.

This could help a bit with rendering performance of the user admin panel too,
i.e. if you accumulate many banned accounts over time.

git-svn-id: http://trac-hacks.org/svn/accountmanagerplugin/trunk@12515 7322e99d-02ea-0310-aa39-e9a107903beb